+++ platform/ui-next/src/components/Popover/index.ts
@@ -1,0 +1,3 @@
+import * as Popover from './Popover';
+
+export default Popover;